DejaVu Sans for LibreOffice systems  and Georgia for Windows only systems added
1. Web-fonts and print-fonts should not be mixed. They have different design
goals.

2. I gave priority have open license fonts. If they are not installed Microsoft
proprietary core fonts are chosen.

3. I based my choice on DejaVu and Liberations fonts. They are [available on all
systems with LibreOffice
installed](http://superuser.com/questions/611583/which-fonts-come-with-openoffice-libreoffice).
*) If not already installed by default all Linux distribution have packages
ready for these fonts.

4. DejaVu is used in all shapes, except for pick-font mono, there I suggest
Liberation mono. It has this "console"-feeling we want and harmonizes well.

5. I left the hyphenate switch on. By default no hyphenation will occur as long
as the user does not install additional hyphenation patterns. But when he does
he does not need to search for the switch to enable hyphenation.